+++ src/resource.rs
@@ -3,7 +3,7 @@ use rrtype;
 use rrtype::RRType;
 use class;
 
-#[deriving(Clone, Show, PartialEq)]
+#[derive(Clone, Debug, PartialEq)]
 pub struct Resource {
   pub name: Vec<String>,
   pub rrtype: RRType,
@@ -13,20 +13,20 @@ pub struct Resource {
   pub rdata: ResourceData
 }
 
-#[deriving(Clone, Show, PartialEq)]
+#[derive(Clone, Debug, PartialEq)]
 pub enum ResourceData {
   A(u8, u8, u8, u8), Raw(Vec<u8>)
 }
 
-pub fn unpack(message: &[u8], offset: uint) -> Option<(Resource, uint)> {
+pub fn unpack(message: &[u8], offset: usize) -> Option<(Resource, usize)> {
   let (name, o) = match name::unpack(message, offset) {
     Some(n) => n, None => return None
   };
 
-  let ty = rrtype::unpack((message[o + 0] as u16 << 8) | (message[o + 1] as u16));
-  let cl = class::unpack((message[o + 2] as u16 << 8) | (message[o + 3] as u16));
-  let ttl = (message[o + 4] as u32 << 24) | (message[o + 5] as u32 << 16) | (message[o + 6] as u32 << 8) | (message[o + 7] as u32);
-  let rlength = (message[o + 8] as u16 << 8) | (message[o + 9] as u16);
+  let ty = rrtype::unpack(((message[o + 0] as u16) << 8) | (message[o + 1] as u16));
+  let cl = class::unpack(((message[o + 2] as u16) << 8) | (message[o + 3] as u16));
+  let ttl = ((message[o + 4] as u32) << 24) | ((message[o + 5] as u32) << 16) | ((message[o + 6] as u32) << 8) | (message[o + 7] as u32);
+  let rlength = ((message[o + 8] as u16) << 8) | (message[o + 9] as u16);
 
   let o = o + 10;
 
@@ -35,13 +35,13 @@ pub fn unpack(message: &[u8], offset: uint) -> Option<(Resource, uint)> {
       ResourceData::A(message[o], message[o + 1], message[o + 2], message[o + 3])
     }
     _ => {
-      let data = message[o .. o+ rlength as uint].to_vec();
+      let data = message[o .. o+ rlength as usize].to_vec();
 
       ResourceData::Raw(data)
     }
   };
 
-  return Some((Resource { name: name, rrtype: ty, class: cl, ttl: ttl, rlength: rlength, rdata: rdata }, o + rlength as uint));
+  return Some((Resource { name: name, rrtype: ty, class: cl, ttl: ttl, rlength: rlength, rdata: rdata }, o + rlength as usize));
 }
 
 #[cfg(test)]
@@ -49,7 +49,7 @@ mod tests {
   use rrtype;
   use class;
 
-  use std::io::File;
+  use std::old_io::File;
 
   #[test]
   fn test_unpack() {
